+++ Catalogo/src/main/java/com/catalogo/IServices/ICategoryService.java
@@ -3,6 +3,6 @@ package com.catalogo.IServices;
 import com.catalogo.domain.Actor;
 import com.catalogo.domain.Category;
 
-public interface ICategoryService extends GenericServiceInterface<Category, Byte>{
+public interface ICategoryService extends GenericServiceInterface<Category, Integer>{
 
 }
